Move InputMap to World & setup overrideable getBodyBounds
